% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faintr_functions.R
\name{compare_cells}
\alias{compare_cells}
\title{Compare means of two subsets of factorial design cells}
\usage{
compare_cells(model, lower_group, higher_group)
}
\arguments{
\item{model}{a brmsfit}

\item{higher}{named list specifying levels of factors that specify
the cell hypothesised to yield a higher dependent variable value}

\item{lower}{named list specifying levels of factors that specify
the cell hypothesised to yield a lower dependent variable value}
}
\value{
a tibble with posterior draws from each of the groups and
  the comparison
}
\description{
This function takes a brms model fit for a factorial design and a
specification of two groups (subsets of design cells) to compare.
A group is specified as a named list, specifiying the factors and
their levels which to include in the group.  It outputs the
posterior samples of the 'higher' and 'lower' subsets of cells,
and the comparison 'higher - lower'.
}
\examples{
m <- brm(pitch ~ gender * context, politeness)
compare_cells(model = m,
              lower = c("gender = M", "context = inf"),
              higher = c("gender = F", "context = pol"))

}
