% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{get_variables}
\alias{get_variables}
\title{Obtaining variable names from a brms model}
\usage{
get_variables(model)
}
\arguments{
\item{model}{Model fit from brms package.}
}
\value{
list with names of the independent variables
}
\description{
For a model for a factorial design, fitted with brms, this function returns the names of the independent variables.
For more information see \code{vignette('faintr_basics')}.
}
\examples{
library(brms)
m <- brm(yield ~ N * P * K, npk)
get_variables(m)
}
\keyword{brms}
\keyword{design,}
\keyword{factorial}
\keyword{regression,}
