% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{make_cell_string}
\alias{make_cell_string}
\title{}
\usage{
make_cell_string(factor_values, factor_info)
}
\arguments{
\item{factor_values}{named list specifying which levels of each factor to combine}

\item{factor_info}{list with names of factors and their levels, including the reference levels (in dummy coding)}
}
\value{
string specifying levels of factors to be combined into a cell
}
\description{
Create string combining factor levels
}
\details{
.. content for \details{} ..
}
