% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politeness.R
\docType{data}
\name{politeness}
\alias{politeness}
\title{Voice pitch depending on context and gender}
\format{
A tibble with 83 rows and 5 columns:
\describe{
  \item{subject}{the speaker}
  \item{gender}{the gender of the speaker}
  \item{sentence}{the identifier of the sentence spoken}
  \item{context}{the context of the sentence, either polite or informal}
  \item{pitch}{the vocal pitch}
}
}
\usage{
politeness
}
\description{
A dataset containg the pitch of voices of speakers of different
genders and contexts.
}
\keyword{datasets}
