% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{compare_cells}
\alias{compare_cells}
\title{Compare cells in factorial regression}
\usage{
compare_cells(model, higher, lower, alpha = 0.05)
}
\arguments{
\item{model}{model of type brmsfit}

\item{higher}{named list specifying levels of factors that specify the cell to give a higher dependent variable value}

\item{lower}{named list specifying levels of factors that specify the cell to give a lower dependent variable value}

\item{alpha}{level of probability}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
